[flexiglass] Fix issues with sceneState/keyguardState, and remove OnSceneAboutToChangeListener.
I'm not sure why sceneState is named that, or why it's type Any?, but it's currently exclusively used to pass a KeyguardState to LockscreenSceneTransitionInteractor, so that we can transiton KTF to that state as part of a scene change to Lockscreen (AOD/DOZING, usually).

OnSceneAboutToChangeListener also seems to only exist to pass this KeyguardState to the LockscreenSceneTransitionInteractor.

This CL clarifies that parameter by explicitly naming it keyguardState and simply passing it to the LSSceneTransitionInteractor when relevant rather than going through a listener that also includes irrelevant scene information (we already use the scene transition information to decide when to use that keyguardState).

Also, fix an issue where this value was dropped on the ground in snapToScene rather than being passed to the listener/LSSceneTransitionInteractor.

Bug: 458502412
Test: atest SystemUITests
Flag: com.android.systemui.scene_container
Change-Id: Ie00d4a1fb267e747083d24e6b64dbf26c9dc9d8a