Create Telecom Service
Provides a mechanism by which the Telecom APEX module
can be started by the framework and provide the
TelecomService implementation.

Provides the TelecomDependencies class, which changes
the src files included in the build based on if
Telecom platform or mainline is being built.

Bug: 445997472
Flag: android.telecom.flags.telecom_mainline_api
Test: Make incoming/outgoing calls; presubmit
Change-Id: I2bd0635161d3beca23dc83936bca55550b322028