Add KEY_APN_MATCHED_REQUIRED carrier config
Introduces a new boolean carrier config key KEY_APN_MATCHED_REQUIRED,
to control whether a successful match in the Android APN database is
mandatory before initiating a data call setup with the modem.

When set to true (the default): Telephony requires a matching APN to be
found in its database. If no match is found for a network request, the
setup data call request will NOT be sent to the modem. This is the
legacy behavior.

When set to false: Telephony will still attempt an APN database lookup.
However, the setup data call request will be sent to the modem
*regardless* of the lookup outcome. This allows the modem to utilize
its own mechanisms, such as URSP (UE Route Selection Policy) rules, to
select the appropriate APN and establish the connection. This value
should be used when the carrier has deployed URSP rules.

This key provides flexibility for carriers transitioning to URSP-based
APN selection.

Test: atest FrameworksTelephonyTests.
Test: permanent failure test, APN-less test.
Test: basic call data test.
Bug: 417753705
Flag: com.android.internal.telephony.flags.enable_traffic_descriptor_connection_capability

Change-Id: I43b7c2d0bb7a7eb74c1c4af2852c75eaf27796a1