Handle Custom Save Location Error Cases
If the screenshot process did not successfully execute when given a custom save location, then we will instead save the screenshot to the default screenshots folder (aka the original process) and inform the user of this behavior through a notification (text provided by a UX writer). An example of when this would happen would be if the user deleted their custom folder and still tried to save to it.

Bug: 444283401
Test: ImageExporterTest
Flag: com.android.systemui.large_screen_screenshot_save_location
Change-Id: Ic236576dda1ae1208af2e2ce62481250178dfb2c