[2/n] Add more flags in recreateOnConfigChanges attribute
This CL adds touchscreen, keyboard, keyboardHidden, navigation, and
colorMode in recreateOnConfigChanges.

After this CL, by default, we don't recreate the activity when these
configs change even if they are not set in configChanges. If the app
wants the activity to be recreated, it needs to set the corresponding
flag in the recreateOnConfigChanges attribute.

Bug: 454795633
Bug: 427186215
Test: manaual tests by connecting / disconnecting keyboard, moving apps
between internal display and connected displays
Flag: com.android.window.flags.enable_less_activity_recreation_on_config_change

Change-Id: I0224f3418d36664a45f81cd813f4f0f0ed36684a