Make ThemeOverlayHelper a real class
Previously, ThemeOverlayHelper was just a collection of static functions. We have changed it into a proper class that can be created and passed around.
This makes it much easier to test ThemeStateManager because we can now give it a fake ThemeOverlayHelper during tests, instead of it always using the real one.

Bug: 333694176
Test: atest FrameworksServicesTests_theme
Flag: android.server.enable_theme_service
Change-Id: I871d2cd9c608deaf82726322e5b51f4a78bb6a09