[res] Set the xml res limit to 1GB, no memcpy for it
The analysis of the APKs from Play shows that some have large
XML resource files, with a few even over 500MB. Thus we have to
support the resources of up to a gigabyte in size.

This CL ensures that we do it optimally - instead of allocating
a second buffer that we copy the content into before parsing, we
move it into the XML parser class, making sure there's no wasted
memory allocations and CPU cycles. This allows the limit of 1GB
to actually only require 1GB of RAM, instead of 2 previously

Test: atest libandroidfw_tests libandroidfw_benchmark \
  idmap2_tests
Flag: android.content.res.xml_file_size_limit
Bug: 443273051
Change-Id: I4f97e8e9ed265087a9964a4a3309e7a6c2f3371b