Make deprecated KTF state uses in Flexiglass throw
As seen in the previous CL as well as e.g. ag/35266450 devs keep adding
deprecated uses of KTF states. This is almost always an error and a
potential bug - if not it has to be at least flag guarded. To prevent
future mistakes like this we now throw so these problems are caught by
the dev immediately and is not overlooked in Logcat.

Since usually flows are setup at startup this kind of error should be
apparent on device start up and there is only small risk of crashing
later on - however this would still be a bug that needs fixing.

Bug: 330311871
Test: Made sure the device is not crashing
Flag: com.android.systemui.scene_container
Change-Id: I87ad8e8e97213db8a9f56948e6fc57d0495eb10e