Add TableLogBuffer for notification alpha values
This change logs values from direct contributing flows to keyguardAlpha (alphaForQsAndShadeExpansion, alphaForTransitions, alphaForTransitionsAndShade, isKeyguardNotVisible), with alphaForTransitions values also broken down by which source flow the alpha came from; also glanceableHubAlpha.

Alpha changes are only logged if they are a change to or from 0 or 1, or (in the case of alphaForTransitions[andShade] specifically) if the alpha comes from a different source, to avoid having a ton of logs during the gradual change from 0-1 but still allow for seeing when changes occur.

Bug: 452888658
Test: view on abt, SharedNotificationContainerViewModelTest
Flag: EXEMPT DEBUG
Change-Id: Ia3eb1a52f73ec65dacdc1bffadb66d694a8512cf