[flexiglass] Fixes bug where UDFPS location was off after boot
See attached bug for discussion and investigation. Long story short: the
UDFPS location might not be ready right after boot so there's a need to
observe an event that tells us that it's ready and then update the
location accordingly.

Fix: 450278802
Test: added a 10 second delay in AuthController#start right before it
calls handleAllFingerprintAuthenticatorsRegistered and rebooted the
device. The lock indicator is shown in the wrong location until 10
seconds later, when it upates to the correct position. Removed that 10
second delay (if was for testing only)
Flag: com.android.systemui.scene_container

Change-Id: Iee5d13e3d2c4502d369efc06e3e89e7a02bb9d0b