Add mic noise cancellation button to audio details view
The existing ANC button is to cancel the environmental noise going to the user's ears. This new mic noise cancellation button is to cancel the noise recorded by the mic.

This CL only creates a skeleton of the implementation. The actual UI will be added in follow up CLs.

Bug: b/460536046
Test: atest
Flag: com.android.systemui.expanded_audio_detailed_view
Change-Id: I3b619da690a306df4fdb6a171d7fe8f74f1d3ac8