Extract visual state to FooterTextButtonModel
When reopening the shade, a new FooterTextButtonViewModel is created with the same values except the onClick lambda. This triggers the crossfade animation even if all other visual data is unchanged.
Extracting the visual data to a separate data class enables us to use transitions on the data that should retrigger animations and ignore the onClick lamdas.

Flag: EXEMPT BUGFIX
Fixes: 457360063
Test: manually - expanding shade with foreground services active
Change-Id: Ic7ef2df8807371bce71271107d2a770db3957c1c