Fix Perfetto to support multiple flows per track event
The previous implementation of PerfettoTrackEventExtra used a single
object for flows and terminating flows. This meant that if multiple
flows were added to the same track event, only the last one would be
recorded.

This change replaces the single Flow objects with a pool of objects,
allowing multiple flows to be associated with a single track event.

The builder methods have been renamed from `setFlow`/`setTerminatingFlow`
to `addFlow`/`addTerminatingFlow` to reflect this new capability, and
the call site in PerfettoSdkTracer has been updated.

Test: TracePerfTest
Bug: 460413832
Flag: NONE - debug api change
Change-Id: I3fdcb532e3d20956ff1dc640a83c8dd1df811969