Introduce InputMethodInfoSafeList
With this CL we start using InputMethodInfoSafeList to transfer
List<InputMethodInfo> over Binder IPC so that we do not need to worry
about TransactionTooLargeException any more. The new behavior is fully
flag-guarded.

This should enable us to remove our past workarounds [1][2] in the
future.

 [1]: Ibb2940fcc02f3b3b51ba6bbe127d646fd7de7c45
      98d4e18d6316c18a3aa24c49b0f6aaed394bc720
 [2]: I51703563414192ee778f30ab57390da1c1a5ded5
      e18d7ba7d8fdfcfdf0d5faef34bf5520daacd8ac

Note on backport:
This change was initially added with a flag guard, but for backporting,
the flag guard is removed and change is directly applied.

Bug: 339761278
Bug: 449416164
Bug: 449181366
Bug: 449393786
Bug: 449227003
Test: atest FrameworksInputMethodSystemServerTests:InputMethodInfoSafeListTest
(cherry picked from commit 949d74963802ea669870331b43c3365fb1919f33)

Change-Id: I0a7667070fcdf17d34b248a5988c38064588718a
Merged-In: I0a7667070fcdf17d34b248a5988c38064588718a