Add activity and task overlays for apps locked by App Lock
This change introduces WindowManager protection for existing tasks and
activities as part of the App Lock feature.

When a locked app is in the background, WindowManagerService now adds a
locked overlay, provided by LockedAppActivity, to the app's tasks and
activities. This prevents the app's content from being visible when the
app is brought to the foreground, until the user authenticates.

If the overlay fails to launch, the task or activity is finished as a
security measure.

Bug: 451912349
Bug: 451913532
Flag: android.security.app_lock_core
Test: manual
Test: atest AppLockOverlayControllerTests
Test: atest TaskTests
Test: atest WindowManagerServiceTests
Change-Id: I46167ad177995241eabf20bdff1f68e92e27464c