Add privileged permission to manage AiSeal virtual machine
Add new permission for preinstalled system apps to access the
AiSealManager API. This permission allows apps to query and manage the
lifecycle of the AiSeal virtual machine and connect to services inside
it.

This permission is just one of the layers preventing access to the
AiSeal API. Additionally, an app needs to be a tenant of AiSeal VM to be
able to use these APIs.

Bug: 455594108
Test: TH, ABTD
Flag: android.aiseal.aiseal_host_apis
Change-Id: Ie8d1ff4762085346b6b422e108cabf0ad1117ce9