Multi-desk support for app sharing
Currently, if we have multi-desks and each desk has more than 2 opened
apps, when we click app sharing, the recent app lsit will only show 2
out of N apps for each desk.

The reason is: the loadRecentTasks in
RecentTaskListProvider only checks taskInfo1 (1st of mTasks) and taskInfo2 (2nd of mTasks) for each
groupedTaskInfo, which doesn't hold for the case when a groupedTaskInfo
is of type desk and has more than 2 apps.

This fix proposes to iterate all tasks for each desk and include them in
the recentTasks.

Bug: 442653998
Bug: 442628274
Flag: com.android.systemui.large_screen_sharing
Test: Manual
Change-Id: I99480e0628d6fe68898609b52d5f647a681588b1