Don't wait for IME to hide when collapsing or expanding bubbles
This change immediately collapses or expands the bubbles stack when
the IME is visible. Previously we were waiting for the IME to hide
before collapsing or expanding in order to avoid flickers, but this
is no longer required due to changes in bubbles and IME.

With the flag on, this also fixes the case where the swipe up gesture
does not work due to bubbles incorrectly stealing it.

Bug: 444859270
Flag: com.android.wm.shell.fix_bubble_swipe_up_gesture
Test: atest BubbleStackViewTest
Test: manual
       - launch Settings in full screen and show IME
       - lock the screen
       - send a bubble
       - tap on the bubble notification in the lock screen
       - observe that the bubble expands and swipe up gesture works
Change-Id: I1b2f86df165c5a011946cb5632792c307212bc23