[flexiglass] Fix bouncer falsing for various intentional actions
Clicking on the user switcher dropdown in bouncer or on the
emergency button will now emit a intentional user input to the
falsing manager. This fixes a bug where clicking on the user
switcher is dismissing a bouncer since the falsingManager rejected
the action.

Bug: 456690090
Test: manual tap on user switcher dropdown and emergency button in
bouncer, observe no falsing increase
Flag: com.android.systemui.scene_container

Change-Id: I2dffba55ad64b18a1e53689b411262c0537cdec8