Clarify transaction error message.
From discussion it seems that this error happens for any error. The same error string is also used elsewhere. Because of this, we often misroute bugs to frozen apps callback. So, I hope to clarify this here so it's clear this is not a frozen apps issue.

Bug: 425390015
Test: N/A
Flag: EXEMPT BUGFIX
Change-Id: I5a236192c67c82e303c7fdc4c21140f15f148761