[flexi] Only read unfoldTranslationX value in layout or draw phase.
`unfoldTranslationXForStartSide` will change every frame when unfolding
the device, so it should *not* be read during composition and should
instead only be read during the layout or draw phases.

This was a regression introduced by always composing the shade. If the
shade is NOT always composed, then `SplitShade` isn't composed on the
screen off -> unfold to lockscreen transition, so the value was never
read. But now that the shade IS always composed, `SplitShade` will be
composed and will read this value.

Bug: 460820693
Flag: com.android.systemui.scene_container
Test: Have shade open while device is unfolded. Slowly fold device and
verify that the shade elements shift in slightly (aka
I91db7922e824ee106faa96709b069f5b862aa7de still applies)
Test: Verify SplitShade does not recompose every frame during the unfold
animation (see trace in bug)

Change-Id: I31a0ce6a08e0ae1c01455de4bcd01bb849122957