[flexiglass] Fixes "stuck" PIN bouncer button highlights.
A change can have a move and an up at the same time. If it's a move, we
consume it before we check if it's an up - but that method returns false
if the change was already consumed, so we never see the up and we never
clear the indication, resulting in the bug.

The fix is to simply ignore the consumption when checking if it's an up.

Fix: 461769518
Test: manually reproduced the bug before the change, couldn't do it
after the change.
Flag: com.android.systemui.scene_container

Change-Id: I4e1da9a5ad6430e79ba8e4463f3c196d790ccc25