Set up for bringing ChronometerState up to par with Chronometer widget
The Chronometer view supports systemTimeMillis-based chronometers, different formatters, and adaptive ticking. They are needed for the full range of Notification.Metric.TimeDifference values.

This is the prework for that -- it changes the API of ChronometerState so that those things can be added with less disruption in future CLs (for example, by having EventTime.EventTimeInstant in addition to EventTime.ElapsedRealtime) -- and changes all the related plumbing. Hopefully the follow up CLs can be only new functionality plus their new tests, without all this brouhaha.

NO_IFTTT=Changes are already in Chronometer

Bug: 444195986
Test: ChronometerStateTest + manually (recording chip, etc)
Flag: EXEMPT PURE_REFACTOR
Change-Id: I23e1c06c114c168843a667d6050b2c6432a15edd