adds package_name to InitAppScanReported
Adds a new field, package_name to the InitAppScanReported metric, which
represents the name of the package being scanned during the boot time
app scan. This is needed in order to break down various dimensions of
the metric, such as signature version by the app name. As part of Full
Stack Integrity (FSI) rollout, we need to ensure v4 signature is fully
rolled out for a given app before enabling strict-signature-checking for
the app (to avoid performance regressions). This new field helps us
monitor the rollout of v4 signature based on the app name, which is not
currently possible using the existing fields.

The InitAppScanReported metric is only emitted for updated system apps.

Bug: 462112193
Flag: EXEMPT METRICS
Test: statsd_testdrive -e 1171
Change-Id: Ic5a308ff11cd1a29e9c635514808777122f18615