Split-screen is only affected by tasks on the same display
We should exit split-screen if a non-split fullscreen task comes over
the top on the same display. The check for this was a little bit too
general because it was looking for any non-split fullscreen task at all
which includes other unrelated displays.

The case where this matters is if we move one of the split tasks to
another display. This creates a CHANGE event on the other display as the
moved Task goes fullscreen. Meanwhile, we should move the other task in
the split to fullscreen as it lost its split pair partner.

Bug: 460447267
Test: Coming in the followup (needs test infra change)
Flag: com.android.window.flags.exit_split_on_display_move_bugfix
Change-Id: I2fdd85b8ad0e8f2ee177d0cb0f87f49e7f7b5935